move PlaceInfo into data submodule
This data structure is shared between indexer and tokenizer.
