+++ nominatim/data/place_info.py
