+++ test/python/tokenizer/sanitizers/test_clean_housenumbers.py
@@ -10,7 +10,7 @@ Tests for the sanitizer that normalizes housenumbers.
 import pytest
 
 from nominatim.tokenizer.place_sanitizer import PlaceSanitizer
-from nominatim.indexer.place_info import PlaceInfo
+from nominatim.data.place_info import PlaceInfo
 
 @pytest.fixture
 def sanitize(request):
