+++ test/python/tokenizer/sanitizers/test_clean_postcodes.py
@@ -10,7 +10,7 @@ Tests for the sanitizer that normalizes postcodes.
 import pytest
 
 from nominatim.tokenizer.place_sanitizer import PlaceSanitizer
-from nominatim.indexer.place_info import PlaceInfo
+from nominatim.data.place_info import PlaceInfo
 from nominatim.tools import country_info
 
 @pytest.fixture
